refactor: Move instantiate and skolemize to base
